TUSC-248 and TUSC-270: Fix widget padding and purchase modal spacing (#319)
Fix widget padding and purchase modal spacing

TUSC-248:

This removes an extra level of padding in the subscriptions widget that
was added as part of the Patternfly 6 migration. Visit /widget-test to
view these changes, or in a deployed environment view it on the
dashboard.

TUSC-270:

This fixes some spacing issues in the purchase modal that appeared
during the Patternfly 6 migration. This just adds padding to the grid
component which is the main content in the modal, and then removes the
margin-right that was on the modal as its handled by the inner padding
now.