TUSC-167: Add error processing to hooks (#330)
* Add error processing to hooks

This addresses unhandled potential errors from API requests - generally,
errors were already "handled" in the components, but not being thrown by
the hooks themselves - and therefore not handled.

This adds thrown errors where needed, as well as adds error handling to
the subscription widget, which was not previously checking for API
errors.

* update test cases